Move web3 connection-related hooks into new folder
