+++ src/components/ConnectWallet/ConnectWalletDialog.tsx
@@ -5,8 +5,8 @@ import Button from 'components/Button'
 import Column from 'components/Column'
 import { Header } from 'components/Dialog'
 import Row from 'components/Row'
-import { Web3Context, Web3ContextType } from 'hooks/useActiveWeb3React'
-import { connections, useConnect, Web3Connection } from 'hooks/useConnectWallet/useProvider'
+import { Web3Context, Web3ContextType } from 'hooks/connectWeb3/useActiveWeb3React'
+import { connections, useConnect, Web3Connection } from 'hooks/connectWeb3/useProvider'
 import styled from 'styled-components/macro'
 import { ThemedText } from 'theme'
