+++ src/components/Swap/SwapButton/index.tsx
@@ -1,10 +1,10 @@
 import { Trans } from '@lingui/macro'
+import useActiveWeb3React from 'hooks/connectWeb3/useActiveWeb3React'
 import { useSwapInfo } from 'hooks/swap'
 import { useSwapApprovalOptimizedTrade } from 'hooks/swap/useSwapApproval'
 import { useSwapCallback } from 'hooks/swap/useSwapCallback'
 import useWrapCallback, { WrapType } from 'hooks/swap/useWrapCallback'
 import { useAddTransaction } from 'hooks/transactions'
-import useActiveWeb3React from 'hooks/useActiveWeb3React'
 import { useSetOldestValidBlock } from 'hooks/useIsValidBlock'
 import useTransactionDeadline from 'hooks/useTransactionDeadline'
 import { Spinner } from 'icons'
