+++ src/components/Swap/index.tsx
@@ -2,11 +2,11 @@ import { JsonRpcProvider } from '@ethersproject/providers'
 import { Trans } from '@lingui/macro'
 import { Provider as Eip1193Provider } from '@web3-react/types'
 import ConnectWallet from 'components/ConnectWallet'
+import useActiveWeb3React from 'hooks/connectWeb3/useActiveWeb3React'
 import { SwapInfoProvider } from 'hooks/swap/useSwapInfo'
 import useSyncConvenienceFee, { FeeOptions } from 'hooks/swap/useSyncConvenienceFee'
 import useSyncTokenDefaults, { TokenDefaults } from 'hooks/swap/useSyncTokenDefaults'
 import { usePendingTransactions } from 'hooks/transactions'
-import useActiveWeb3React from 'hooks/useActiveWeb3React'
 import useHasFocus from 'hooks/useHasFocus'
 import useOnSupportedNetwork from 'hooks/useOnSupportedNetwork'
 import { useAtom } from 'jotai'
