+++ src/hooks/multicall.ts
@@ -1,4 +1,4 @@
-import useActiveWeb3React from 'hooks/useActiveWeb3React'
+import useActiveWeb3React from 'hooks/connectWeb3/useActiveWeb3React'
 import useBlockNumber from 'hooks/useBlockNumber'
 import multicall from 'state/multicall'
