+++ src/hooks/routing/useClientSideSmartOrderRouterTrade.ts
@@ -9,8 +9,8 @@ import { useCallback, useMemo } from 'react'
 import { GetQuoteResult, InterfaceTrade, TradeState } from 'state/routing/types'
 import { computeRoutes, transformRoutesToTrade } from 'state/routing/utils'
 
+import useActiveWeb3React from '../connectWeb3/useActiveWeb3React'
 import useWrapCallback, { WrapType } from '../swap/useWrapCallback'
-import useActiveWeb3React from '../useActiveWeb3React'
 import { useGetIsValidBlock } from '../useIsValidBlock'
 import usePoll from '../usePoll'
 import { useRoutingAPIArguments } from './useRoutingAPIArguments'
