+++ src/hooks/useAutoSlippageTolerance.ts
@@ -2,7 +2,7 @@ import { Trade } from '@uniswap/router-sdk'
 import { Currency, CurrencyAmount, Percent, TradeType } from '@uniswap/sdk-core'
 import { SUPPORTED_GAS_ESTIMATE_CHAIN_IDS } from 'constants/chains'
 import { L2_CHAIN_IDS } from 'constants/chains'
-import useActiveWeb3React from 'hooks/useActiveWeb3React'
+import useActiveWeb3React from 'hooks/connectWeb3/useActiveWeb3React'
 import useNativeCurrency from 'hooks/useNativeCurrency'
 import JSBI from 'jsbi'
 import { useMemo } from 'react'
