+++ src/hooks/useClientSideV3Trade.ts
@@ -1,8 +1,8 @@
 import { Currency, CurrencyAmount, TradeType } from '@uniswap/sdk-core'
 import { Route, SwapQuoter } from '@uniswap/v3-sdk'
 import { SupportedChainId } from 'constants/chains'
+import useActiveWeb3React from 'hooks/connectWeb3/useActiveWeb3React'
 import { useSingleContractWithCallData } from 'hooks/multicall'
-import useActiveWeb3React from 'hooks/useActiveWeb3React'
 import JSBI from 'jsbi'
 import { useMemo } from 'react'
 import { InterfaceTrade, TradeState } from 'state/routing/types'
