+++ src/hooks/useCurrency.ts
@@ -2,8 +2,8 @@ import { arrayify } from '@ethersproject/bytes'
 import { parseBytes32String } from '@ethersproject/strings'
 import { Currency, Token } from '@uniswap/sdk-core'
 import { TOKEN_SHORTHANDS } from 'constants/tokens'
+import useActiveWeb3React from 'hooks/connectWeb3/useActiveWeb3React'
 import { NEVER_RELOAD, useSingleCallResult } from 'hooks/multicall'
-import useActiveWeb3React from 'hooks/useActiveWeb3React'
 import { useBytes32TokenContract, useTokenContract } from 'hooks/useContract'
 import useNativeCurrency from 'hooks/useNativeCurrency'
 import { useMemo } from 'react'
