+++ src/hooks/useCurrencyBalance.ts
@@ -3,8 +3,8 @@ import { Currency, CurrencyAmount, Token } from '@uniswap/sdk-core'
 import ERC20ABI from 'abis/erc20.json'
 import { Erc20Interface } from 'abis/types/Erc20'
 import { nativeOnChain } from 'constants/tokens'
+import useActiveWeb3React from 'hooks/connectWeb3/useActiveWeb3React'
 import { useMultipleContractSingleData, useSingleContractMultipleData } from 'hooks/multicall'
-import useActiveWeb3React from 'hooks/useActiveWeb3React'
 import { useInterfaceMulticall } from 'hooks/useContract'
 import JSBI from 'jsbi'
 import { useMemo } from 'react'
