+++ src/hooks/useERC20Permit.ts
@@ -6,8 +6,8 @@ import { Trade as V2Trade } from '@uniswap/v2-sdk'
 import { Trade as V3Trade } from '@uniswap/v3-sdk'
 import { SWAP_ROUTER_ADDRESSES, V3_ROUTER_ADDRESS } from 'constants/addresses'
 import { DAI, UNI, USDC_MAINNET } from 'constants/tokens'
+import useActiveWeb3React from 'hooks/connectWeb3/useActiveWeb3React'
 import { useSingleCallResult } from 'hooks/multicall'
-import useActiveWeb3React from 'hooks/useActiveWeb3React'
 import JSBI from 'jsbi'
 import { useMemo, useState } from 'react'
