+++ src/hooks/useNativeCurrency.ts
@@ -1,7 +1,7 @@
 import { NativeCurrency } from '@uniswap/sdk-core'
 import { SupportedChainId } from 'constants/chains'
 import { nativeOnChain } from 'constants/tokens'
-import useActiveWeb3React from 'hooks/useActiveWeb3React'
+import useActiveWeb3React from 'hooks/connectWeb3/useActiveWeb3React'
 import { useMemo } from 'react'
 
 export default function useNativeCurrency(): NativeCurrency {
