+++ src/hooks/useOnSupportedNetwork.ts
@@ -1,7 +1,7 @@
 import { ALL_SUPPORTED_CHAIN_IDS } from 'constants/chains'
 import { useMemo } from 'react'
 
-import useActiveWeb3React from './useActiveWeb3React'
+import useActiveWeb3React from './connectWeb3/useActiveWeb3React'
 
 function useOnSupportedNetwork() {
   const { chainId } = useActiveWeb3React()
