+++ src/hooks/usePools.ts
@@ -4,8 +4,8 @@ import { abi as IUniswapV3PoolStateABI } from '@uniswap/v3-core/artifacts/contra
 import { computePoolAddress } from '@uniswap/v3-sdk'
 import { FeeAmount, Pool } from '@uniswap/v3-sdk'
 import { V3_CORE_FACTORY_ADDRESSES } from 'constants/addresses'
+import useActiveWeb3React from 'hooks/connectWeb3/useActiveWeb3React'
 import { useMultipleContractSingleData } from 'hooks/multicall'
-import useActiveWeb3React from 'hooks/useActiveWeb3React'
 import JSBI from 'jsbi'
 import { useMemo } from 'react'
 import { IUniswapV3PoolStateInterface } from 'types/v3/IUniswapV3PoolState'
