+++ src/hooks/useTokenList/useQueryTokens.ts
@@ -1,5 +1,5 @@
 import { nativeOnChain } from 'constants/tokens'
-import useActiveWeb3React from 'hooks/useActiveWeb3React'
+import useActiveWeb3React from 'hooks/connectWeb3/useActiveWeb3React'
 import { useTokenBalances } from 'hooks/useCurrencyBalance'
 import useDebounce from 'hooks/useDebounce'
 import { useMemo } from 'react'
