+++ src/hooks/useTransactionDeadline.ts
@@ -6,7 +6,7 @@ import { useAtom } from 'jotai'
 import { useMemo } from 'react'
 import { transactionTtlAtom } from 'state/settings'
 
-import useActiveWeb3React from './useActiveWeb3React'
+import useActiveWeb3React from './connectWeb3/useActiveWeb3React'
 
 /** Returns the default transaction TTL for the chain, in minutes. */
 export function useDefaultTransactionTtl(): number {
