+++ src/state/multicall.tsx
@@ -1,5 +1,5 @@
 import { createMulticall } from '@uniswap/redux-multicall'
-import useActiveWeb3React from 'hooks/useActiveWeb3React'
+import useActiveWeb3React from 'hooks/connectWeb3/useActiveWeb3React'
 import useBlockNumber from 'hooks/useBlockNumber'
 import { useInterfaceMulticall } from 'hooks/useContract'
 import { combineReducers, createStore } from 'redux'
